Authentication adn Authoriztion

+++ controller/api/v1/posts_api.js
@@ -20,11 +20,17 @@ module.exports.index = async function (req, res) {
 module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
-    post.remove();
-    await Comments.deleteMany({ post: req.params.id });
-    return res.json(200, {
-      message: "Post and associated comments deleted sucessfully!",
-    });
+    if (post.user == req.user.id) {
+      post.remove();
+      await Comments.deleteMany({ post: req.params.id });
+      return res.json(200, {
+        message: "Post and associated comments deleted sucessfully!",
+      });
+    } else {
+      return res.json(401, {
+        message: "You can not delete this post",
+      });
+    }
   } catch (err) {
     console.log(err);
     return res.json(500, {
